backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
isort==6.0.1
phonenumbers==9.0.9
pillow==11.3.0
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
